backend/requirements.txt
# requirements.txt для бэкенда
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
PyJWT==2.8.0
email-validator==2.1.0